@@ -14,23 +14,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "4.0.0",
-    "@angular/compiler": "4.0.0",
-    "@angular/core": "4.0.0",
-    "@angular/forms": "4.0.0",
-    "@angular/http": "4.0.0",
-    "@angular/platform-browser": "4.0.0",
-    "@angular/platform-browser-dynamic": "4.0.0",
-    "@angular/router": "4.0.0",
+    "@angular/animations": "4.3.3",
+    "@angular/common": "4.3.3",
+    "@angular/compiler": "4.3.3",
+    "@angular/core": "4.3.3",
+    "@angular/forms": "4.3.3",
+    "@angular/platform-browser": "4.3.3",
+    "@angular/platform-browser-dynamic": "4.3.3",
+    "@angular/router": "4.3.3",
     "admin-lte": "2.3.11",
     "core-js": "2.4.1",
     "font-awesome": "4.7.0",
     "intl": "1.2.5",
-    "jquery": "3.4.0",
-    "reflect-metadata": "0.1.9",
-    "rxjs": "5.1.0",
+    "jquery": "3.1.1",
+    "reflect-metadata": "0.1.10",
+    "rxjs": "5.4.2",
     "ts-helpers": "1.1.2",
-    "zone.js": "0.8.4"
+    "web-animations-js": "2.2.5",
+    "zone.js": "0.8.16"
   },
   "devDependencies": {
     "@angular/cli": "1.2.7",
