@@ -1,13 +1,13 @@
-import {Response} from '@angular/http';
+import {HttpErrorResponse} from '@angular/common/http';
 import {Observable} from 'rxjs';
 
 export class ErrorHandler {
 
-  static handlerError(error: Response|any) {
+  static handlerError(error: HttpErrorResponse|any) {
     let errorMessage: string;
 
-    if (error instanceof  Response) {
-      errorMessage = `Erro ${error.status} ao acessar a URL ${error.url} - ${error.statusText}`
+    if (error instanceof HttpErrorResponse) {
+      errorMessage = `Erro ${error.status} ao acessar a URL ${error.url} - ${error.statusText} - BODY ${error.error}`
     } else {
       errorMessage = error.toString()
     }
