@@ -3,7 +3,8 @@ constructor(public address: string,
             public number: number,
             public optionalAddress: string,
             public paymentOption: string,
-            public orderItems: OrderItem[] = []) { }
+            public orderItems: OrderItem[] = [],
+            public id?: string) { }
 }
 
 class OrderItem {
