@@ -4,16 +4,14 @@ import {CartItem} from '../restaurant-detail/shopping-cart/cart-item.model';
 import {Order} from './order.model';
 import {Observable} from 'rxjs';
 import 'rxjs/add/operator/map';
-import {Headers, Http, RequestOptions} from '@angular/http';
+import { HttpClient} from '@angular/common/http';
 import {MEAT_API} from '../app.api';
 
 @Injectable()
 export class OrderService {
 
   constructor(private cartService: ShoppingCartService,
-              private http: Http) {
-
-  }
+              private http: HttpClient) { }
 
   cartItems(): CartItem[] {
     return this.cartService.items
@@ -40,11 +38,7 @@ export class OrderService {
   }
 
   checkOrder(order: Order): Observable<string> {
-    const headers = new Headers();
-    headers.append('Content-Type', 'application/json');
-
-    return this.http.post(`${MEAT_API}/orders`, JSON.stringify(order), new RequestOptions({headers}))
-      .map(response => response.json())
+    return this.http.post<Order>(`${MEAT_API}/orders`, order)
       .map(order => order.id);
   }
 }
