@@ -1,6 +1,6 @@
 import {Restaurant} from './restaurant/restaurant.model';
 import {Injectable} from '@angular/core';
-import {Http} from '@angular/http';
+import {HttpClient, HttpParams} from '@angular/common/http';
 import {MEAT_API} from '../app.api';
 import {Observable} from 'rxjs';
 import 'rxjs/add/operator/map';
@@ -11,29 +11,27 @@ import {MenuItem} from '../restaurant-detail/menu-item/menu-item.model';
 @Injectable()
 export class RestaurantsService {
 
-  constructor(private http: Http) {}
+  constructor(private http: HttpClient) {}
 
-  restaurants(): Observable<Restaurant[]> {
-    return this.http.get(`${MEAT_API}/restaurants`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handlerError);
+  restaurants(search?: string): Observable<Restaurant[]> {
+    let params: HttpParams = undefined;
+
+    if (search) {
+      params = new HttpParams().set('q', search);
+    }
+
+    return this.http.get<Restaurant[]>(`${MEAT_API}/restaurants`, {params});
   }
 
   restaurantById(id: string): Observable<Restaurant> {
-    return this.http.get(`${MEAT_API}/restaurants/${id}`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handlerError);
+    return this.http.get<Restaurant>(`${MEAT_API}/restaurants/${id}`);
   }
 
-  reviewsOfRestaurant(id: string): Observable<Restaurant> {
-    return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handlerError);
+  reviewsOfRestaurant(id: string): Observable<any> {
+    return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`);
   }
 
   menuOfRestaurant(id: string): Observable<MenuItem[]> {
-    return this.http.get(`${MEAT_API}/restaurants/${id}/menu`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handlerError);
+    return this.http.get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`);
   }
 }
